fix: fix a lot of configurations to integrate with front end

+++ src/main/java/br/edu/ifsudeste/demo/model/service/MarcaService.java
@@ -1,10 +1,13 @@
 package br.edu.ifsudeste.demo.model.service;
 
 import br.edu.ifsudeste.demo.exception.RegraNegocioException;
+import br.edu.ifsudeste.demo.model.entity.Dispositivo;
 import br.edu.ifsudeste.demo.model.entity.Marca;
 import br.edu.ifsudeste.demo.model.entity.Modelo;
+import br.edu.ifsudeste.demo.model.entity.Produto;
 import br.edu.ifsudeste.demo.model.repository.MarcaRepository;
 import br.edu.ifsudeste.demo.model.repository.ModeloRepository;
+import br.edu.ifsudeste.demo.model.repository.ProdutoRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
 
@@ -18,10 +21,12 @@ public class MarcaService {
 
     private final MarcaRepository repository;
     private final ModeloRepository modeloRepository;
+    private final ProdutoRepository produtoRepository;
 
-    public MarcaService(MarcaRepository repository, ModeloRepository modeloRepository) {
+    public MarcaService(MarcaRepository repository, ModeloRepository modeloRepository,  ProdutoRepository produtoRepository) {
         this.repository = repository;
         this.modeloRepository = modeloRepository;
+        this.produtoRepository = produtoRepository;
     }
 
     public List<Marca> getMarca() {
@@ -48,6 +53,14 @@ public class MarcaService {
                     .map(Modelo::getNomeModelo)
                     .collect(Collectors.joining(", "));
             throw new RegraNegocioException("Falha ao excluir: Marca sendo utilizada no(s) modelo(s) " + nomesModelos);
+        }
+
+        List<Produto> produtosUtilizando = produtoRepository.findByMarca(marca);
+        if (!produtosUtilizando.isEmpty()) {
+            String nomesProdutos = produtosUtilizando.stream()
+                    .map(Produto::getNome)
+                    .collect(Collectors.joining(", "));
+            throw new RegraNegocioException("Falha ao excluir: Marca sendo utilizada no(s) produto(s) " + nomesProdutos);
         }
 
         repository.delete(marca);
